requirements.txt
jwt
flask_mail
flask_restful
flask_couchdb
azure-cognitiveservices-speech
gunicorn